Untangle dependencies between tools.execute and tools.actions.
